app/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
pydantic[email]
passlib
argon2-cffi
python-jose[cryptography]
python-dotenv
requests
os
transformers


